pyproject.toml
[project]
name = "dicom-search"
version = "0.1.0"
description = "DICOM mammogram classification pipeline with active learning"
requires-python = ">=3.10"
dependencies = [
    "fastai>=2.7",
    "torch>=2.0",
    "python-fasthtml>=0.4",
    "sqlite-utils>=3.35",
    "pydicom>=2.4",
    "joblib>=1.3",
    "onnxruntime-gpu>=1.16",
    "onnx>=1.15",
    "onnxscript>=0.1",
    "pillow>=10.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "tqdm>=4.66",
    "ipython>=8.0",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = []
